src_execute_sql/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.0.0",
]
